+++ weblate/addons/scripts.py
@@ -52,17 +52,29 @@ class BaseScriptAddon(BaseAddon):
         self.execute_process(component, command, environment)
         self.trigger_alerts(component)
 
-    def post_push(self, component) -> None:
+    def post_push(self, component, activity_log_id: int | None = None) -> None:
         self.run_script(component)
 
-    def post_update(self, component, previous_head: str, skip_push: bool) -> None:
+    def post_update(
+        self,
+        component,
+        previous_head: str,
+        skip_push: bool,
+        activity_log_id: int | None = None,
+    ) -> None:
         self.run_script(component, env={"WL_PREVIOUS_HEAD": previous_head})
 
-    def post_commit(self, component: Component, store_hash: bool) -> None:
+    def post_commit(
+        self, component: Component, store_hash: bool, activity_log_id: int | None = None
+    ) -> None:
         self.run_script(component=component)
 
     def pre_commit(
-        self, translation: Translation, author: str, store_hash: bool
+        self,
+        translation: Translation,
+        author: str,
+        store_hash: bool,
+        activity_log_id: int | None = None,
     ) -> None:
         self.run_script(translation=translation)
 
@@ -73,5 +85,5 @@ class BaseScriptAddon(BaseAddon):
             )
             translation.addon_commit_files.append(filename)
 
-    def post_add(self, translation) -> None:
+    def post_add(self, translation, activity_log_id: int | None = None) -> None:
         self.run_script(translation=translation)